+++ frontend/src/components/laundries/listItem.jsx
@@ -2,12 +2,9 @@ import React, {useState} from "react";
 import {laundryImage} from "../../lib/common";
 import Loading from '../common/loading.jsx'
 import Modal from "react-modal"
-import Button from "../common/button";
-import {Link} from "react-router-dom";
-import {deleteLaundry, getLaundry, washed} from "../../lib/api/laundries";
-import {useHistory} from "react-router";
-import {createLaundryHistories} from "../../lib/api/laundry_histories";
+import {getLaundry} from "../../lib/api/laundries";
 import ModalContents from "./modalContents";
+import ModalButtons from "./modalButtons";
 
 const customStyles = {
     content: {
@@ -25,8 +22,6 @@ Modal.setAppElement("#root")
 const ListItem = ({id, name, image}) => {
     const [modalIsOpen, setIsOpen] = useState(false)
     const [isLoading, setIsLoading] = useState(false)
-    const [isShownDeleteMode, setIsShownDeleteMode] = useState(false)
-    const history = useHistory()
 
     function openModal() {
         setIsOpen(true);
@@ -49,77 +44,6 @@ const ListItem = ({id, name, image}) => {
             setLaundry(res.data.data)
         } catch (err) {
             console.error(err)
-        }
-    }
-
-    const deleteLaundryInfo = async (e, id) => {
-        e.preventDefault()
-
-        try {
-            const res = await deleteLaundry(id)
-            setIsShownDeleteMode(false)
-            console.log(res)
-            history.push("/laundries")
-        } catch (err) {
-            console.error(err)
-        }
-    }
-
-    // 洗濯するボタン
-    const handleWash = async (e, laundryId) => {
-        e.preventDefault()
-
-        // 洗濯履歴を作る
-        try {
-            const res = await createLaundryHistories(laundryId)
-            console.log(res)
-        } catch (err) {
-            console.error(err)
-        }
-
-        // 洗濯物のwashAtを更新、今日は非表示にする
-        try {
-            const res = await washed(laundryId)
-            console.log(res)
-        } catch (err) {
-            console.error(err)
-        }
-    }
-
-    let buttons
-    if (isShownDeleteMode) {
-        buttons = () => {
-            return (
-                <>
-                    <p className="text-center text-red-500">本当に削除しますか？</p>
-
-                    <div className="mt-8 w-2/3 mx-auto flex justify-between space-x-2">
-                        <Button color="pink" func={(e) => {
-                            deleteLaundryInfo(e, laundry.id)
-                        }} value="はい"/>
-                        <Button color="gray" func={() => {
-                            setIsShownDeleteMode(false)
-                        }} value="いいえ"/>
-                    </div>
-                </>
-            )
-        }
-    } else {
-        buttons = () => {
-            return (
-                <div className="mt-8 flex justify-around text-sm">
-                    <Button color="pink" func={(e) => {
-                        handleWash(e, id)
-                    }} value="今日洗濯する"/>
-                    <Link to={`/laundries/${laundry.id}`}
-                          className="bg-pink-100 hover:bg-pink-200 py-2 px-3 rounded-xl border-b-4 border-pink-500">
-                        変更する
-                    </Link>
-                    <Button color="gray" func={() => {
-                        setIsShownDeleteMode(true)
-                    }} value="削除する"/>
-                </div>
-            )
         }
     }
 
@@ -148,7 +72,7 @@ const ListItem = ({id, name, image}) => {
                         description={laundry.washAt}
                     />
 
-                    {buttons()}
+                    <ModalButtons id={laundry.id}/>
 
                 </Loading>
             </Modal>
